+++ themes/theme-docs/src/tokens.ts
@@ -34,6 +34,20 @@ const blue = {
   950: '#172554',
 };
 
+const green = {
+  50: '#f0fdf4',
+  100: '#ecfccb',
+  200: '#d9f99d',
+  300: '#bef264',
+  400: '#a3e635',
+  500: '#84cc16',
+  600: '#65a30d',
+  700: '#4d7c0f',
+  800: '#3f6212',
+  900: '#365314',
+  950: '#1a2e05',
+};
+
 const neutral = {
   50: '#fafafa',
   100: '#f5f5f5',
@@ -46,6 +60,20 @@ const neutral = {
   800: '#262626',
   900: '#171717',
   950: '#0a0a0a',
+};
+
+const red = {
+  50: '#fef2f2',
+  100: '#fee2e2',
+  200: '#fecaca',
+  300: '#fca5a5',
+  400: '#f87171',
+  500: '#ef4444',
+  600: '#dc2626',
+  700: '#b91c1c',
+  800: '#991b1b',
+  900: '#7f1d1d',
+  950: '#450a0a',
 };
 
 const slate = {
@@ -132,7 +160,9 @@ export const colors = {
     },
 
     // Status
-    info: blue[100],
+    success: green[100],
+    error: red[50],
+    info: blue[50],
     warning: amber[50],
   },
 
@@ -141,5 +171,9 @@ export const colors = {
   border: {
     DEFAULT: slate[300],
     primary: slate[950],
+
+    // status
+    success: green[600],
+    error: red[600],
   },
 };